reena_chatbot/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-groq
